Dynamic viscosity
Added dynamic viscosity to the atmosphere class to make it easier and clearer to compute the partials of kinematic viscosity with respect to altitude

Kinematic viscosity partials

Implemented the partials for the kinematic viscosity. Also made some slight change to the way the pressure is computed to reduce the time it takes to compute by using np functions instead of the python power

Speed of sound partials

Implemented the partial for speed of sound. Also made a slight change to the way the speed of sound is computed as, in the particular cas of the square root, np.sqrt is way faster than **0.5

Density partial

Density, test with altitude in m and test with altitude in ft

Pressure partials

Computation of the derivatives of pressure with respect to the altitude.
Also for the test, I decided, in the end to add proper fixture for atmosphere in feet because the difference is significant so it is worth testing